Refactor all libraries to interfaces Fix time sync issues

+++ src/ws2812/ws2812_patterns.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-typedef void (*pattern_fn)(uint32_t *stream, size_t len, uint32_t beat_pos);
+typedef void (*pattern_fn)(uint32_t *stream, size_t len, uint8_t beat_pos);
 
 typedef struct {
   pattern_fn pat;
@@ -17,6 +17,6 @@ size_t get_pattern_count();
 const char *pattern_get_name(uint8_t pattern_idx);
 
 void run_pattern(int pattern_idx, uint32_t *stream, size_t len,
-                 uint32_t beat_pos);
+                 uint8_t beat_pos);
 
 #endif // WS2812_PATTERNS_H
